requirements.txt
binance-futures-connector
numpy
pandas
requests
python-dotenv
pymysql
FastAPI